Extract base class for exceptions
